+++ __tests__/schema/uuid/abstract-repository.ts
@@ -21,14 +21,11 @@ import {
   license,
   page,
   pageRevision,
-  solution,
-  solutionRevision,
   user,
   video,
   videoRevision,
 } from '../../../__fixtures__'
 import { nextUuid, getTypenameAndId, given, Client } from '../../__utils__'
-import { getDefaultLicense } from '~/config/licenses'
 import { Model } from '~/internals/graphql'
 import {
   EntityRevisionType,
@@ -88,11 +85,6 @@ const repositoryFixtures: Record<
     repository: groupedExercise,
     revision: groupedExerciseRevision,
     revisionType: EntityRevisionType.GroupedExerciseRevision,
-  },
-  [EntityType.Solution]: {
-    repository: solution,
-    revision: solutionRevision,
-    revisionType: EntityRevisionType.SolutionRevision,
   },
   [EntityType.Video]: {
     repository: video,
@@ -232,13 +224,6 @@ describe('Repository', () => {
                 ... on AbstractRepository {
                   license {
                     id
-                    instance
-                    default
-                    title
-                    shortTitle
-                    url
-                    content
-                    agreement
                   }
                 }
               }
@@ -247,32 +232,6 @@ describe('Repository', () => {
         })
         .withVariables({ id: repository.id })
         .shouldReturnData({ uuid: { license } })
-    },
-  )
-
-  test.each(repositoryCases)(
-    '%s uses default license if licenseId is not supported',
-    async (_type, { repository }) => {
-      given('UuidQuery').for({ ...repository, licenseId: 42 })
-
-      await new Client({ userId: user.id })
-        .prepareQuery({
-          query: gql`
-            query license($id: Int!) {
-              uuid(id: $id) {
-                ... on AbstractRepository {
-                  license {
-                    id
-                  }
-                }
-              }
-            }
-          `,
-        })
-        .withVariables({ id: repository.id })
-        .shouldReturnData({
-          uuid: { license: { id: getDefaultLicense(repository.instance).id } },
-        })
     },
   )