+++ scripts/changelog.ts
@@ -1604,6 +1604,31 @@ async function exec(): Promise<void> {
           'Return arbitrary JSON instead of a string from content generation service',
         ],
       },
+      {
+        tagName: 'v0.57.9',
+        date: '2023-11-10',
+        fixed: [
+          'Return `null` when event cannot be loaded',
+          'enmeshed: Do not set name outside of user journey',
+        ],
+      },
+      {
+        tagName: 'v0.57.10',
+        date: '2023-11-10',
+        fixed: ['enmeshed: Do not set surname and given name on init'],
+      },
+      {
+        tagName: 'v0.57.11',
+        date: '2023-11-11',
+        internal: ['Remove solutions'],
+        fixed: ['enmeshed: Store cache only for 20min'],
+      },
+      {
+        tagName: 'v0.58.0',
+        date: '2023-11-13',
+        changed: ['remove license queries, simplify `License` type'],
+        internal: ['Move content generation service logic to api'],
+      },
     ],
   })